75%core done: i just figured out how to use context and it rocks

+++ fseps/filesystem/fseps.py
@@ -3,13 +3,14 @@ import time
 import errno
 import datetime
 
-from fuse import FuseOSError, Operations
+from fuse import FuseOSError, Operations, fuse_get_context
+from settings import FSEPS_PERMISSIONS_FILE_NAME
 
 
 class FsEps(Operations):
     def __init__(self, root):
         self.root = root
-
+        print self.root
     # Helpers
     # =======
     def _full_path(self, partial):
@@ -27,6 +28,7 @@ class FsEps(Operations):
     # ==================
 
     def access(self, path, mode):
+        print type(path), path, 'xxxxx>', type(mode), mode,'<<<'
         full_path = self._full_path(path)
         if not os.access(full_path, mode):
             raise FuseOSError(errno.EACCES)
@@ -43,23 +45,29 @@ class FsEps(Operations):
         full_path = self._full_path(path)
         st = os.lstat(full_path)
 
+        # print dir(st)
 
+        print '**************'
         x = dict((key, getattr(st, key)) for key in ('st_atime', 'st_ctime',
                                                         'st_gid', 'st_mode',
                                                         'st_mtime', 'st_nlink',
-                                                        'st_size', 'st_uid'))
+                                                        'st_size', 'st_uid', 'st_rdev'))
         print "-------------------->"
+        # x['st_uid'] = 1001
         for k, v in x.items():
-            print str(k) + "->"+ str(type(v)) + str(v)
+            print str(k) + "->" + str(v)
         return x
 
     def readdir(self, path, fh):
-
+        print '@@@@@@@@@@@@@@@@@@@@@@@@@@'
+        print fuse_get_context()
+        print '@@@@@@@@@@@@@@@@@@@@@@@@@@'
         full_path = self._full_path(path)
 
         dirents = ['.', '..']
         if os.path.isdir(full_path):
             dirents.extend(os.listdir(full_path))
+        dirents = [x for x in dirents if x != FSEPS_PERMISSIONS_FILE_NAME]
         for r in dirents:
             yield r
 
@@ -72,6 +80,9 @@ class FsEps(Operations):
             return pathname
 
     def mknod(self, path, mode, dev):
+        print "create file "
+        print path, mode, dev
+        print "done\n"
         return os.mknod(self._full_path(path), mode, dev)
 
     def rmdir(self, path):
@@ -82,17 +93,22 @@ class FsEps(Operations):
         print "mkdir->"
         return os.mkdir(self._full_path(path), mode)
 
-    # def statfs(self, path):
-    #     full_path = self._full_path(path)
-    #     stv = os.statvfs(full_path)
-    #     return dict((key, getattr(stv, key)) for key in ('f_bavail', 'f_bfree',
-    #                                                      'f_blocks', 'f_bsize',
-    #                                                      'f_favail', 'f_ffree',
-    #                                                      'f_files', 'f_flag',
-    #                                                      'f_frsize',
-    #                                                      'f_namemax'))
+    def statfs(self, path):
+        print '\n\n statfs$$$$$$$$$$$$\n\n'
+        full_path = self._full_path(path)
+        stv = os.statvfs(full_path)
+        return dict((key, getattr(stv, key)) for key in ('f_bavail', 'f_bfree',
+                                                         'f_blocks', 'f_bsize',
+                                                         'f_favail', 'f_ffree',
+                                                         'f_files', 'f_flag',
+                                                         'f_frsize',
+                                                         'f_namemax'))
 
     def unlink(self, path):
+        dir_path, filename = os.path.split(path)
+        if filename == FSEPS_PERMISSIONS_FILE_NAME:
+            raise FuseOSError(errno.ENOENT)
+        print "DELETED SOMETHING"
         return os.unlink(self._full_path(path))
 
     def symlink(self, target, name):
@@ -109,6 +125,9 @@ class FsEps(Operations):
 
     # File methods
     def open(self, path, flags):
+        dir_path, filename = os.path.split(path)
+        if filename == FSEPS_PERMISSIONS_FILE_NAME:
+            raise FuseOSError(errno.ENOENT)
         upper_bound = datetime.time(18, 0, 0)
         lower_bound = datetime.time(9, 0, 0)
         now = datetime.datetime.now().time()
@@ -118,6 +137,10 @@ class FsEps(Operations):
         return os.open(full_path, flags)
 
     def create(self, path, mode, fi=None):
+        print "file created "
+        dir_path, filename = os.path.split(path)
+        if filename == FSEPS_PERMISSIONS_FILE_NAME:
+            raise FuseOSError(errno.ENOENT)
         full_path = self._full_path(path)
         return os.open(full_path, os.O_WRONLY | os.O_CREAT, mode)
 
@@ -141,5 +164,4 @@ class FsEps(Operations):
         return os.close(fh)
 
     def fsync(self, path, fdatasync, fh):
-        time.sleep(10)
         return self.flush(path, fh)
